+++ src/Concerns/CoreBindings.php
@@ -9,6 +9,8 @@ use Monolog\Handler\StreamHandler;
 use Monolog\Formatter\LineFormatter;
 use Illuminate\Filesystem\Filesystem;
 use Laravel\Lumen\Routing\UrlGenerator;
+use Zend\Diactoros\Response as PsrResponse;
+use Symfony\Bridge\PsrHttpMessage\Factory\DiactorosFactory;
 
 trait CoreBindings
 {
@@ -67,6 +69,8 @@ trait CoreBindings
         'Illuminate\Contracts\Queue\Queue'                => 'registerQueueBindings',
         'redis'                                           => 'registerRedisBindings',
         'request'                                         => 'registerRequestBindings',
+        'Psr\Http\Message\ServerRequestInterface'         => 'registerPsrRequestBindings',
+        'Psr\Http\Message\ResponseInterface'              => 'registerPsrResponseBindings',
         'Illuminate\Http\Request'                         => 'registerRequestBindings',
         'session'                                         => 'registerSessionBindings',
         'session.store'                                   => 'registerSessionBindings',
@@ -128,6 +132,7 @@ trait CoreBindings
             'request'                                         => 'Illuminate\Http\Request',
             'Illuminate\Session\SessionManager'               => 'session',
             'Illuminate\Session\Store'                        => 'session.store',
+            'Laravel\Lumen\Routing\UrlGenerator'              => 'url',
             'Illuminate\Contracts\View\Factory'               => 'view',
         ];
     }
@@ -399,6 +404,30 @@ trait CoreBindings
             $this->configure('services');
 
             return $this->loadComponent('mail', 'Illuminate\Mail\MailServiceProvider', 'mailer');
+        });
+    }
+
+    /**
+     * Register container bindings for the PSR-7 request implementation.
+     *
+     * @return void
+     */
+    protected function registerPsrRequestBindings()
+    {
+        $this->singleton('Psr\Http\Message\ServerRequestInterface', function () {
+            return (new DiactorosFactory())->createRequest($this->make('request'));
+        });
+    }
+
+    /**
+     * Register container bindings for the PSR-7 response implementation.
+     *
+     * @return void
+     */
+    protected function registerPsrResponseBindings()
+    {
+        $this->singleton('Psr\Http\Message\ResponseInterface', function () {
+            return new PsrResponse();
         });
     }